+++ tests/test_problem.py
@@ -25,7 +25,7 @@ import unittest
 import numpy as np
 import scipy.sparse as spa
 
-from qpsolvers import Problem, ProblemError
+from qpsolvers import ActiveSet, Problem, ProblemError
 
 from .problems import get_sd3310_problem
 
@@ -64,22 +64,32 @@ class TestProblem(unittest.TestCase):
             Problem(P, q, G, h, None, b).check_constraints()
 
     def test_cond(self):
-        self.assertGreater(self.problem.cond(), 200.0)
+        active_set = ActiveSet(
+            G_indices=range(self.problem.G.shape[0]),
+            lb_indices=[],
+            ub_indices=[],
+        )
+        self.assertGreater(self.problem.cond(active_set), 200.0)
 
     def test_cond_unconstrained(self):
         unconstrained = Problem(self.problem.P, self.problem.q)
-        self.assertAlmostEqual(unconstrained.cond(), 124.257, places=4)
+        active_set = ActiveSet()
+        self.assertAlmostEqual(
+            unconstrained.cond(active_set), 124.257, places=4
+        )
 
     def test_cond_no_equality(self):
         no_equality = Problem(
             self.problem.P, self.problem.q, self.problem.G, self.problem.h
         )
-        self.assertGreater(no_equality.cond(), 200.0)
+        active_set = ActiveSet(G_indices=range(self.problem.G.shape[0]))
+        self.assertGreater(no_equality.cond(active_set), 200.0)
 
     def test_cond_sparse(self):
         sparse = Problem(spa.csc_matrix(self.problem.P), self.problem.q)
+        active_set = ActiveSet()
         with self.assertRaises(ProblemError):
-            sparse.cond()
+            sparse.cond(active_set)
 
     def test_check_matrix_shapes(self):
         Problem(np.eye(1), np.ones(1))